app/tests/requirements.txt
requests==2.29.0
pytest==7.3.1
grpcio==1.51.3
grpcio-tools==1.51.3